moving FramedAgentType to frame module; stub aggregate as frame param
